feat(FR-1514): implement remote artifact list view on the Reservoir Page
